requirements.txt
flask
deepgram-sdk
flask-socketio
numpy
sounddevice
load_dotenv
flask-cors